[build_wheels] Support building from existing sdist.

+++ scripts/build_wheels.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 # This is for running in a manylinux2010 docker image, so /bin/bash is fine.
-# docker run --rm -v /some/where:/out:rw -v /path/to/accelerator:/accelerator:ro --tmpfs /tmp:exec,size=1G quay.io/pypa/manylinux2010_x86_64 /accelerator/scripts/build_wheels.sh 20xx.xx.xx.dev1 commit/tag/branch
+# docker run --rm -v /some/where:/out:rw -v /path/to/accelerator:/accelerator:ro --tmpfs /tmp:exec,size=1G quay.io/pypa/manylinux2010_x86_64 /accelerator/scripts/build_wheels.sh 20xx.xx.xx.dev1 [commit/tag/branch]
+# (builds sdist if you specify a commit, needs sdist to already be in wheelhouse otherwise)
 
 set -euo pipefail
 
-if [ "$#" != "2" ]; then
-	echo "Usage: $0 ACCELERATOR_BUILD_VERSION commit/tag/branch"
+if [ "$#" != 1 -a "$#" != "2" ]; then
+	echo "Usage: $0 ACCELERATOR_BUILD_VERSION [commit/tag/branch]"
 	exit 1
 fi
 
@@ -28,15 +29,20 @@ esac
 VERSION="$1"
 NAME="accelerator-$(echo "$1" | sed s/\\.0/./g)"
 
-cd /tmp
-rm -rf accelerator
-git clone /accelerator
-cd accelerator
-git checkout $2
-ACCELERATOR_BUILD_VERSION="$VERSION" /opt/python/cp38-cp38/bin/python3 ./setup.py sdist
-cp dist/"$NAME".tar.gz /out/wheelhouse/
-cd ..
-rm -rf accelerator
+if [ "$#" = "1" ]; then
+	test -e /out/wheelhouse/"$NAME".tar.gz || exit 1
+else
+	test -e /out/wheelhouse/"$NAME".tar.gz && exit 1
+	cd /tmp
+	rm -rf accelerator
+	git clone /accelerator
+	cd accelerator
+	git checkout $2
+	ACCELERATOR_BUILD_VERSION="$VERSION" /opt/python/cp38-cp38/bin/python3 ./setup.py sdist
+	cp dist/"$NAME".tar.gz /out/wheelhouse/
+	cd ..
+	rm -rf accelerator
+fi
 
 
 MANYLINUX_VERSION="${AUDITWHEEL_PLAT/%_*}"
